Refactoring Content module.
  - Views pulled apart by context.
  - Models pulled apart by context.
  - Fixed the trailing slash bug on urls
  - Fixes issue #42
