+++ .travis.yml
@@ -1,11 +1,10 @@
 sudo: true
-before_install:
-  - sudo apt-get update -qq
-  - sudo apt-get install -qq build-essential gettext python-dev zlib1g-dev libpq-dev xvfb
-  - sudo apt-get install -qq libtiff4-dev libjpeg8-dev libfreetype6-dev liblcms1-dev libwebp-dev
-  - sudo apt-get install -qq graphviz-dev python-setuptools python3-dev python-virtualenv python-pip
-  - sudo apt-get install -qq firefox automake libtool libreadline6 libreadline6-dev libreadline-dev
-  - sudo apt-get install -qq libsqlite3-dev libxml2 libxml2-dev libssl-dev libbz2-dev wget curl llvm
+install:
+  - pip install -r requirements/base.txt
+  - pip install -r requirements/local.txt
+  - pip install -r requirements/test.txt
 language: python
 python:
-  - "3.5"
+  - "3.6"
+script:
+  - pytest # or py.test for Python versions 3.5 and below
